requirements.txt
azure-core==1.23.0
azure-storage-blob==12.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
dlib @ https://github.com/jloh02/dlib/releases/download/v19.22/dlib-19.22.99-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
idna==3.3
isodate==0.6.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
Pillow==9.0.1
pycparser==2.21
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
